PM-1239: Removed dependency on pegasus-config. Improved tests.

+++ test/unit/transfer/test.sh
@@ -170,20 +170,14 @@ EOF
 
 export TEST_DIR=`pwd`
 
-cd ../../../bin
-rm -f pegasus-transfer pegasus-integrity
-ln -s pegasus-python-wrapper pegasus-transfer
-ln -s pegasus-python-wrapper pegasus-integrity
+if [ "x$1" = "x" ]; then
+    echo "Please specify the Pegasus bin dir as the first argument" >&2
+    exit 1
+fi
+export PATH=$1:$PATH
 
-export PATH=`pwd`:$PATH
-export TRANSFER_LOCATION=`pwd`/pegasus-transfer
-
-cd $TEST_DIR
-
-export KICKSTART_LOCATION=`cd ../../../src/tools/pegasus-kickstart && pwd`/pegasus-kickstart
-
-# we require kickstart
-(cd ../../../src/tools/pegasus-kickstart && make) >/dev/null 2>&1
+export TRANSFER_LOCATION=$1/pegasus-transfer
+export KICKSTART_LOCATION=$1/pegasus-kickstart
 
 export KICKSTART_INTEGRITY_DATA=ks.integrity.$$
 rm -f $KICKSTART_INTEGRITY_DATA
